tools/biblio-validator/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "biblio-validator"
version = "1.0.0"
description = "Professional citation validation tool for academic documents"
readme = "README.md"
license = "MIT"
authors = [
    {name = "Deep Biblio Tools Team", email = "team@deepbiblio.tools"}
]
keywords = ["citations", "validation", "bibliography", "academic", "crossref", "pubmed"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Text Processing :: Linguistic",
]
requires-python = ">=3.8"
dependencies = [
    "click>=8.0",
    "requests>=2.28",
    "bibtexparser>=1.4",
    "pydantic>=2.0",
    "rich>=13.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.0",
]

[project.scripts]
biblio-validator = "biblio_validator.cli:main"

[project.urls]
Homepage = "https://github.com/deep-biblio-tools/deep-biblio-tools"
Documentation = "https://deep-biblio-tools.readthedocs.io/tools/biblio-validator"
"Bug Tracker" = "https://github.com/deep-biblio-tools/deep-biblio-tools/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/biblio_validator"]
